src/compute-manager/pyproject.toml
[tool.poetry]
name = "compute-manager"
version = "0.1.0"
description = "A lightweight compute manager with pluggable drivers"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
pyaml = ">=25.1.0,<26.0.0"
jsonschema = ">=4.23.0,<5.0.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
pydantic = ">=2.11.4,<3.0.0"
docker = "^7.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
